staticx: Refactor generic 'hooks' sub-package
