+++ staticx/api.py
@@ -3,7 +3,7 @@
 # https://github.com/JonathonReinhart/staticx
 #
 import shutil
-from tempfile import NamedTemporaryFile, mkdtemp
+from tempfile import NamedTemporaryFile
 import os
 import logging
 from itertools import chain
@@ -13,65 +13,7 @@ from .utils import *
 from .elf import *
 from .archive import SxArchive
 from .constants import *
-
-
-def process_pyinstaller_archive(ar, prog):
-    # See utils/cliutils/archive_viewer.py
-
-    # If PyInstaller is not installed, do nothing
-    try:
-        from PyInstaller.archive.readers import CArchiveReader, NotAnArchiveError
-    except ImportError:
-        return
-
-    # Attempt to open the program as PyInstaller archive
-    try:
-        pyi_ar = CArchiveReader(prog)
-    except:
-        # Silence all PyInstaller exceptions here
-        return
-    logging.info("Opened PyInstaller archive!")
-
-    # Create a temporary directory
-    # TODO PY3: Use tempfile.TemporaryDirectory and cleanup()
-    tmpdir = mkdtemp(prefix='staticx-pyi-')
-
-    try:
-        # Process the archive, looking at all shared libs
-        for n, item in enumerate(pyi_ar.toc.data):
-            (dpos, dlen, ulen, flag, typcd, name) = item
-
-            # Only process binary files
-            # See xformdict in PyInstaller.building.api.PKG
-            if typcd != 'b':
-                continue
-
-            # Extract it to a temporary location
-            x, data = pyi_ar.extract(n)
-            tmppath = os.path.join(tmpdir, name)
-            logging.debug("Extracting to {}".format(tmppath))
-            with open(tmppath, 'wb') as f:
-                f.write(data)
-
-            # Silence "you do not have execution permission" warning from ldd
-            make_executable(tmppath)
-
-            # Add any missing libraries to our archive
-            for libpath in get_shobj_deps(tmppath):
-                lib = os.path.basename(libpath)
-
-                if lib in ar.libraries:
-                    logging.debug("{} already in staticx archive".format(lib))
-                    continue
-
-                if pyi_ar.toc.find(lib) != -1:
-                    logging.debug("{} already in pyinstaller archive".format(lib))
-                    continue
-
-                ar.add_library(libpath)
-    finally:
-        shutil.rmtree(tmpdir)
-
+from .hooks import run_hooks
 
 def generate_archive(prog, interp, extra_libs=None):
     logging.info("Program interpreter: " + interp)
@@ -89,8 +31,7 @@ def generate_archive(prog, interp, extra_libs=None):
         for libpath in chain(get_shobj_deps(prog), extra_libs):
             ar.add_library(libpath)
 
-
-        process_pyinstaller_archive(ar, prog)
+        run_hooks(ar, prog)
 
     f.flush()
     return f
